Restaurant Menu Management
A python program to manage the menu of a restaurant to take order, add or remove item to the menu, calculate bill based on the ordered item. You can also update the price of any item when necessary. Customers can see the menu which are currently available and place an order. The items are stored in a csv file with their respective price.